Ease up the slow file test, for pypy
The concurrency test is essentially a race between a file read,
with a 1 second delay, and 40 1-byte net reads, with a 10msec delay.

Pypy can fail this test, for reasons that appear to have nothing to
do with the goal of the test.

Ease the goal to allow pypy to catch up.